Create Exercise 2
Problem: You are given multidimentional array with "M" rows and "N" columns. Make an algorithm which creates new array as every "i" element in this array is equal to the average of all negative elements in the "i" row of the multidimentional array.

Shows basic knowledge of working and navigating with arrays and multidimentional arrays.